+++ cmd/local/jvmcollect/jfr.go
@@ -27,10 +27,19 @@ import (
 
 func RunCollectJFR(c *conf.CollectConf) error {
 	var w bytes.Buffer
+	w = bytes.Buffer{}
 	if err := ddcio.Shell(&w, fmt.Sprintf("jcmd %v VM.unlock_commercial_features", c.DremioPID())); err != nil {
 		simplelog.Warningf("Error trying to unlock commercial features %v. Note: newer versions of OpenJDK do not support the call VM.unlock_commercial_features. This is usually safe to ignore", err)
 	}
 	simplelog.Debugf("node: %v - jfr unlock commercial output - %v", c.NodeName(), w.String())
+
+	w = bytes.Buffer{}
+	if err := ddcio.Shell(&w, fmt.Sprintf("jcmd %v JFR.stop name=\"DREMIO_JFR\"", c.DremioPID())); err != nil {
+		simplelog.Debugf("attempting to stop existing JFR failed, but this is usually expected: '%v' -- output: '%v'", err, w.String())
+	} else {
+		simplelog.Warningf("JFR named DREMIO_JFR was running on PID %v and has been stopped so we could get a fresh collection: '%v'", c.DremioPID(), w.String())
+	}
+
 	w = bytes.Buffer{}
 	if err := ddcio.Shell(&w, fmt.Sprintf("jcmd %v JFR.start name=\"DREMIO_JFR\" settings=profile maxage=%vs  filename=%v/%v.jfr dumponexit=true", c.DremioPID(), c.DremioJFRTimeSeconds(), c.JFROutDir(), c.NodeName())); err != nil {
 		return fmt.Errorf("unable to run JFR due to error %v", err)