+++ pkg/simplelog/logger_test.go
@@ -20,6 +20,8 @@ import (
 	"fmt"
 	"strings"
 	"testing"
+
+	"github.com/dremio/dremio-diagnostic-collector/pkg/output"
 )
 
 func TestLogger(t *testing.T) {
@@ -67,6 +69,41 @@ func TestLogger(t *testing.T) {
 		}
 	}
 }
+
+func TestStartLogMessage(t *testing.T) {
+	InitLogger(4)
+	loc := GetLogLoc()
+	if loc == "" {
+		t.Error("expected log file to not be empty but it was")
+	}
+	out, err := output.CaptureOutput(func() {
+		LogStartMessage()
+	})
+	if err != nil {
+		t.Fatalf("failed running capture %v", err)
+	}
+	if !strings.Contains(out, loc) {
+		t.Errorf("expected %v in string %v", loc, out)
+	}
+}
+
+func TestEndLogMessage(t *testing.T) {
+	InitLogger(4)
+	loc := GetLogLoc()
+	out, err := output.CaptureOutput(func() {
+		LogEndMessage()
+	})
+	if err != nil {
+		t.Fatalf("failed running capture %v", err)
+	}
+	if loc == "" {
+		t.Error("expected log file to not be empty but it was")
+	}
+	if !strings.Contains(out, loc) {
+		t.Errorf("expected %v in string %v", loc, out)
+	}
+}
+
 func TestLoggerMessageIsTruncated(t *testing.T) {
 	var arr []string
 	for i := 0; i < 2000; i++ {